added manage listings

+++ app/Http/Controllers/ListingController.php
@@ -57,6 +57,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect("/")->with('message', 'Listing created successfully!');
@@ -73,6 +75,11 @@ class ListingController extends Controller
     //Update listing data
     public function update(Request $request, Listing $listing)
     {
+
+        //check if logged in user is the owner of the listing
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action.');
+        }
         $formFields = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -95,7 +102,17 @@ class ListingController extends Controller
     //Delete listing
     public function destroy(Listing $listing)
     {
+        //check if logged in user is the owner of the listing
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action.');
+        }
         $listing->delete();
         return redirect("/")->with('message', 'Listing deleted successfully!');
+    }
+
+    //Manage listings
+    public function manage()
+    {
+        return view('listings.manage', ["listings" => auth()->user()->listings()->get()]);
     }
 }
